backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gedcom-explorer-backend"
version = "1.0.0"
description = "Backend API for GEDCOM Explorer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-gedcom>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
